SimpleWindowBuilder now accepts ActiveEventLoop
- Refactored SimpleWindowBuilder into a module in a separate file
- Added a GliumEventLoop trait similar to
  glutin_winit::GlutinEventLoop
  - Implemented for EventLoop, ActiveEventLoop
  - Required to support the newer non-deprecated ActiveEventLoop
    method of Window creation without introducing breaking changes
    (solution stolen from glutin_winit)
  - Duplication required because the glutin_winit trait is not
    public
